segment number datatype changed from uint16_t to uint32_t

+++ src/patternMatcher.cpp
@@ -71,7 +71,7 @@ public:
 			{
 				throw strus::runtime_error(_TXT("term event orig size out of range"));
 			}
-			else if (term.origseg() >= (std::size_t)std::numeric_limits<uint16_t>::max())
+			else if (term.origseg() >= (std::size_t)std::numeric_limits<uint32_t>::max())
 			{
 				throw strus::runtime_error(_TXT("term event orig segment number out of range"));
 			}
